# [Feat + Fix] Faster calculations and retrieval of data + Fix breaking non-scalable UI