--- a/ui/backend/client_ui.py
+++ b/ui/backend/client_ui.py
@@ -38,6 +38,26 @@ class UiClient(Client):
         result = cursor.fetchone()[0]
         return result
 
+    def get_all_discoveries_count(self):
+        """ Get the total number of discoveries and the number of True
+        Positives per repo
+
+        Returns
+        -------
+        list
+            A list of tuples containing (repo_url, total, true positive).
+        """
+        query = '''SELECT repo_url,
+                          COUNT(*) as total,
+                          sum(case when STATE='new' then 1 else 0 end) as tp 
+                          FROM discoveries 
+                          GROUP BY repo_url;
+                '''
+        cursor = self.db.cursor()
+        cursor.execute(query)
+        result = cursor.fetchall()
+        return result
+
     def get_files_summary(self, query, repo_url):
         """ Get aggregated discoveries info on all files of a repository.