--- a/ui/server.py
+++ b/ui/server.py
@@ -302,12 +302,21 @@ def scan_repo():
 
 @app.route('/get_repos')
 def get_repos():
+    # Get data from the database
     active_scans = _get_active_scans()
-
     repos = c.get_repos()
+    repos_metadata = c.get_all_discoveries_count()
+
     for repo in repos:
-        repo['total'] = c.get_discoveries_count(repo['url'])
-        repo['TP'] = c.get_discoveries_count(repo['url'], state='new')
+        for metadata in repos_metadata:
+            if(repo.get('total')):
+                continue
+            if(repo['url'] == metadata[0]):
+                repo['total'] = metadata[1]  # Total number of discoveries
+                repo['TP'] = metadata[2]  # Total number of true positives
+            else:
+                repo['total'] = 0
+                repo['TP'] = 0
         repo['scan_active'] = False
         if repo['url'] in active_scans:
             repo['scan_active'] = True