### Introduction
When the user scans a lot of repositories, they fill-up the database with multiple new entries. 
When the database grows in size, collecting data becomes a lengthy task and blocks the UI (by keeping the other SQL queries pending) resulting in a non-functional UI (home page).

### Fix
We propose doing the calculations of the metadata with a multi-purpose SQL query that returns the following:

- Number of discoveries per repo
- Number of **true positive** discoveries per repo (part of the total number of the discoveries calculated previously)

### Performance
| Database | Old Time     | New Time       |
|----------|--------------|----------------|
| SQLite   | ~ 14 seconds | ~ 2 seconds    |
| Postgres | ~ 7 seconds  | ~ 0.05 seconds |