@pep8speaks 2021-08-12
Hello @alaabenfatma! Thanks for opening this PR. 

* In the file [`ui/backend/client_ui.py`](https://github.com/SAP/credential-digger/blob/ea401c2dcdf8ed276a8d1eb542f05757eab06c13/ui/backend/client_ui.py):

> [Line 52:77](https://github.com/SAP/credential-digger/blob/ea401c2dcdf8ed276a8d1eb542f05757eab06c13/ui/backend/client_ui.py#L52): [W291](https://duckduckgo.com/?q=pep8%20W291) trailing whitespace
> [Line 53:43](https://github.com/SAP/credential-digger/blob/ea401c2dcdf8ed276a8d1eb542f05757eab06c13/ui/backend/client_ui.py#L53): [W291](https://duckduckgo.com/?q=pep8%20W291) trailing whitespace



Do see the [Hitchhiker's guide to code style](https://goo.gl/hqbW4r)