@marcorosa 2021-08-13
As a suggestion for future PRs/analyses: if you want to compare performance you should give an indication of the setting you're using. In this specific case, for example, you manage to pass from 14 seconds to 2 seconds. But what's the setting? 14s vs 2s it's not a constant time, it strongly depends on the number of repos and discoveries together with the hardware you're running the UI on (a lag when you have 1GB RAM is not the same of a lag when you have 16GB RAM, and I could tell the same for the internet connection in case you use a remote server).

By the way the PR is good and I will merge it soon.